backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.4
supabase==2.5.1
qdrant-client==1.9.1
httpx==0.27.0
google-generativeai==0.7.2
groq==0.9.0
PyMuPDF==1.24.3
pdfplumber==0.11.0
pandas==2.2.2
tiktoken==0.7.0
pytest==8.2.0
pytest-asyncio==0.23.7
respx==0.21.1
